chore: Text 컴포넌트 디자인시스템으로 이동
